+++ src/components/TopBar/components/Nav.tsx
@@ -16,6 +16,7 @@ const Nav: React.FC = () => {
         </Button>
       </SLink>
       <StyledRouterLink to="/help">Help</StyledRouterLink>
+      <StyledRouterLink to="/daoinfo">DAO</StyledRouterLink>
     </StyledNav>
   );
 };