+++ RELEASE-NOTES.md
@@ -58,6 +58,22 @@
 - hotfix for staked token value calculations
 - updated TGE1 farm pair name
 
+## Alpha
+- add farm tier label to the Tentacle farms (tier 1, tier 2, etc...)
+- added tooltip to tier label
+- adjusted dao info page text (summary)
+
+DAO Info Page
+- ecosystemTvl added
+- circulating INK added NOTE: Calculates based on harvested INK. IE, unharvested INK is not circulating based on this calculation.
+- daily transaction count added
+- prices added
+- DAO holdings added, including multisig and farm addresses
+- add mission statement
+- marketcap added
+- ToS added to footer
+
+
 ## Unreleased
 - build version shown in footer
 - links to charts added